plugins.vc.player: autoplay: play or add to queue if the new message contains audio and its duration is less than 10 min

+++ plugins/vc/player.py
@@ -148,24 +148,39 @@ async def playout_ended_handler(group_call, filename):
 # - Pyrogram handlers
 
 
-@Client.on_message(main_filter & current_vc & filters.regex("^(\\/|!)play$"))
+@Client.on_message(
+    filters.group
+    & ~filters.edited
+    & current_vc
+    & (filters.regex("^(\\/|!)play$") | filters.audio)
+)
 async def play_track(client, m: Message):
     group_call = mp.group_call
     playlist = mp.playlist
-    # show playlist
-    if not m.reply_to_message or not m.reply_to_message.audio:
+    # check audio
+    if m.audio:
+        m_audio = m
+    elif m.reply_to_message and m.reply_to_message.audio:
+        m_audio = m.reply_to_message
+    else:
         await mp.send_playlist()
         await m.delete()
         return
+    if m_audio.audio.duration > 600:
+        reply = await m.reply_text(
+            f"{emoji.ROBOT} audio which duration longer than 10 min won't be "
+            "automatically added to playlist"
+        )
+        await _delay_delete_messages((reply, ), DELETE_DELAY)
+        return
     # check already added
-    m_reply = m.reply_to_message
     if playlist and playlist[-1].audio.file_unique_id \
-            == m_reply.audio.file_unique_id:
+            == m_audio.audio.file_unique_id:
         reply = await m.reply_text(f"{emoji.ROBOT} already added")
         await _delay_delete_messages((reply, m), DELETE_DELAY)
         return
     # add to playlist
-    playlist.append(m_reply)
+    playlist.append(m_audio)
     if len(playlist) == 1:
         m_status = await m.reply_text(
             f"{emoji.INBOX_TRAY} downloading and transcoding..."
@@ -182,7 +197,8 @@ async def play_track(client, m: Message):
     await mp.send_playlist()
     for track in playlist[:2]:
         await download_audio(track)
-    await m.delete()
+    if not m.audio:
+        await m.delete()
 
 
 @Client.on_message(main_filter
